SpacebioEngine/requirements.txt
pandas
numpy
tqdm
transformers
sentence-transformers
torch
faiss-cpu
scikit-learn
streamlit
ipywidgets
matplotlib
networkx
joblib
